requirements.txt
django-cors-headers == 3.13.0
django == 4.1.1
django-rest-auth == 0.9.5
djangorestframework-jwt == 1.11.0
djangorestframework == 3.14.0